💚 CI 빌드 수정 : 소나큐브 CI/CD 수정 디스코드 웹훅으로 분석 결과를 받을 수 있게 수정
