+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
     id 'org.springframework.boot' version '3.1.3'
     id 'io.spring.dependency-management' version '1.1.3'
-    id "org.sonarqube" version "4.3.0.3225"
+    id "org.sonarqube" version "6.0.1.5171"
 }
 
 group = 'com.PhotoCardTradeProjectBack'
@@ -134,10 +134,9 @@ clean.doLast {
     file(querydslDir).deleteDir()
 }
 
-sonarqube {
+sonar {
     properties {
-        property "sonar.projectKey", "capo-backend"
-        property "sonar.host.url", "http://13.124.217.23:9000"
-        property "sonar.login", "squ_c4fb1b4048ff7711b6b12c2cc43c45b900b09982"
+        property "sonar.projectKey", "capo-project_capo-backend_89d062f5-c979-4114-b0c5-9f3797033f98"
+        property "sonar.projectName", "capo-backend"
     }
 }
